+++ random_permeability.h
@@ -13,10 +13,18 @@
 #ifndef RANDOM_PERMEABILITY_H
 #define RANDOM_PERMEABILITY_H
 
-#include "darcy.h"
+#include <deal.II/base/function.h>
+#include <deal.II/base/point.h>
+#include <deal.II/base/tensor.h>
+#include <deal.II/base/tensor_function.h>
+
+#include <cmath>
+#include <vector>
 
 namespace darcy
 {
+  using namespace dealii;
+
   namespace RandomMedium
   {
     // =========================================================================
